requirements.txt
python-dotenv
pygame
openai